moved CrossRelationProducer classes into own folder
