+++ src/Propel/Generator/Builder/Om/ObjectBuilder/RelationCodeProducer/CrossRelationNames.php
@@ -6,7 +6,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Propel\Generator\Builder\Om\ObjectBuilder;
+namespace Propel\Generator\Builder\Om\ObjectBuilder\RelationCodeProducer;
 
 use Propel\Generator\Builder\Om\NameProducer;
 use Propel\Generator\Builder\Om\ReferencedClasses;
