+++ src/Propel/Generator/Builder/Om/ObjectBuilder/RelationCodeProducer/FunctionArgumentSignatureCollector.php
@@ -6,7 +6,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Propel\Generator\Builder\Om\ObjectBuilder;
+namespace Propel\Generator\Builder\Om\ObjectBuilder\RelationCodeProducer;
 
 use Propel\Generator\Model\Column;
